chore(typescript): Update session routes to use typescript
